# api.yaml
openapi: 3.0.3
info:
  title: PyHost Cloud API
  version: "1.0.0"
  description: |
    Public developer API and internal Agent API for PyHost Cloud.
    Public endpoints use Bearer tokens (user session / API key).
    Agent endpoints require HMAC signed requests with rotating secrets.
servers:
  - url: https://api.pyhost.example.com
    description: Production
  - url: https://staging-api.pyhost.example.com
    description: Staging
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "User/API token for developer API (server-issued)."
    AgentHMAC:
      type: apiKey
      in: header
      name: X-Agent-ID
      description: >
        Agent authentication requires these headers:
        - X-Agent-ID: agent UUID
        - X-Timestamp: unix epoch seconds
        - X-Request-ID: client-generated UUID for idempotency/tracing
        - X-Signature: HMAC_SHA256(secret, METHOD|PATH|X-Timestamp|BODY_SHA256)
        Provide the secret per-agent; rotate regularly.
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
        details:
          type: object
...
    EnvVar:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
        is_secret:
          type: boolean
...
    AgentHeartbeat:
      type: object
      properties:
        agent_uuid:
          type: string
        region:
          type: string
        capacity:
          type: object
          properties:
            cpu_cores: { type: integer }
            memory_mb: { type: integer }
            disk_mb: { type: integer }
        running_projects:
          type: integer
        timestamp:
          type: string
          format: date-time
paths:
...
  /agent/v1/heartbeat:
    post:
      summary: Agent heartbeat (agent → control plane)
      description: >
        Agent must send heartbeat regularly. Must include headers:
        X-Agent-ID, X-Timestamp, X-Request-ID, X-Signature (HMAC).
      security:
        - AgentHMAC: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentHeartbeat'
      responses:
        '200':
          description: OK
  /agent/v1/projects/{projectId}/deploy:
    post:
      summary: Control plane triggers deploy on agent (control → agent)
      description: >
        Control plane calls agent with deploy instruction. Agent must validate signature and project ownership.
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      security:
        - AgentHMAC: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                deployment_id: { type: integer }
                artifact_url: { type: string }
                start_command: { type: string }
                env:
                  type: array
                  items:
                    $ref: '#/components/schemas/EnvVar'
      responses:
        '200':
          description: Agent accepted deploy
        '400':
          $ref: '#/components/schemas/Error'
  /agent/v1/projects/{projectId}/start:
    post:
      summary: Start a project process (control → agent)
      security:
        - AgentHMAC: []
      responses:
        '200':
          description: accepted
  /agent/v1/projects/{projectId}/stop:
    post:
      summary: Stop a project process (control → agent)
      security:
        - AgentHMAC: []
      responses:
        '200':
          description: accepted
  /agent/v1/projects/{projectId}/status:
    get:
      summary: Agent → control plane / control plane → agent project runtime status endpoint
      security:
        - AgentHMAC: []
      responses:
        '200':
          description: status payload
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  pid: { type: integer }
                  uptime_seconds: { type: number }
                  memory_mb: { type: number }
                  cpu_percent: { type: number }
security:
  - bearerAuth: []  # default for interactive clients; agent endpoints explicitly require AgentHMAC
...